AIBot/pyproject.toml
[project]
name = "fastapi-multi-tenant-ai"
version = "0.1.0"
description = "FastAPI Multi-Tenant Authentication with AI Chat Service"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.0",  # avoid incompatibility causing '__about__' AttributeError with passlib
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-groq>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "groq>=0.4.0",
    "google-generativeai>=0.3.0",
    "tenacity>=8.2.0",
    "httpx>=0.25.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "email-validator>=2.2.0",
    "websockets>=12.0",
]
readme = "README.md"
requires-python = ">= 3.11"


[tool.ruff]
target-version = "py311"
line-length = 88

[tool.pytest.ini_options]
testpaths = ["src/Backend/Test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
